# curso-introduccion-apis/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/productos/estado/{idProducto}":{"put":{"tags":["producto-controller"],"operationId":"cambiarEstadoProducto","parameters":[{"name":"idProducto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string","enum":["DISPONIBLE","NO_DISPONIBLE"]}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/productos/actualizar/{idProducto}":{"put":{"tags":["producto-controller"],"operationId":"actualizarProducto","parameters":[{"name":"idProducto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"Nombre del producto","in":"query","required":true,"schema":{"type":"string"}},{"name":"descripcion","in":"query","required":true,"schema":{"type":"string"}},{"name":"precio","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"cantidad","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"estado","in":"query","required":true,"schema":{"type":"string","enum":["DISPONIBLE","NO_DISPONIBLE"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/categorias/{idCategoria}":{"get":{"tags":["categoria-controller"],"operationId":"obtenerCategoriaPorId","parameters":[{"name":"idCategoria","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Categoria"}}}}}},"put":{"tags":["categoria-controller"],"operationId":"actualizarCategoria","parameters":[{"name":"idCategoria","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Categoria"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Categoria"}}}}}},"delete":{"tags":["categoria-controller"],"operationId":"eliminarCategoria","parameters":[{"name":"idCategoria","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/productos/registrar/{categoriaId}":{"post":{"tags":["producto-controller"],"operationId":"registrarProducto","parameters":[{"name":"categoriaId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"Nombre del producto","in":"query","required":true,"schema":{"type":"string"}},{"name":"descripcion","in":"query","required":true,"schema":{"type":"string"}},{"name":"precio","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"cantidad","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"estado","in":"query","required":true,"schema":{"type":"string","enum":["DISPONIBLE","NO_DISPONIBLE"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/messages":{"get":{"tags":["message-controller"],"operationId":"listaMensajes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Message"}}}}}}},"post":{"tags":["message-controller"],"operationId":"crearMensaje","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Message"}}}}}}},"/api/categorias":{"get":{"tags":["categoria-controller"],"operationId":"listarCategorias","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Categoria"}}}}}}},"post":{"tags":["categoria-controller"],"operationId":"crearCategoria","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Categoria"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Categoria"}}}}}}},"/api/productos":{"get":{"tags":["producto-controller"],"operationId":"listarProducto","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Producto"}}}}}}}},"/api/productos/estado/{estado}":{"get":{"tags":["producto-controller"],"operationId":"listarProductosPorEstado","parameters":[{"name":"estado","in":"path","required":true,"schema":{"type":"string","enum":["DISPONIBLE","NO_DISPONIBLE"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Producto"}}}}}}}},"/api/productos/buscar/nombre/{nombre}":{"get":{"tags":["producto-controller"],"operationId":"buscarPorNombre","parameters":[{"name":"nombre","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/productos/buscar/id/{idProducto}":{"get":{"tags":["producto-controller"],"operationId":"bucarPorId","parameters":[{"name":"idProducto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/messages/{id}":{"get":{"tags":["message-controller"],"operationId":"obtenerMessagePorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Message"}}}}}},"delete":{"tags":["message-controller"],"operationId":"eliminarMensaje","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/productos/{idProducto}":{"delete":{"tags":["producto-controller"],"operationId":"eliminarProducto","parameters":[{"name":"idProducto","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"Categoria":{"type":"object","properties":{"idCategoria":{"type":"integer","format":"int64"},"nombreCategoria":{"type":"string"},"categoria":{"$ref":"#/components/schemas/Categoria"}}},"Message":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"content":{"type":"string"}}},"Producto":{"type":"object","properties":{"idProducto":{"type":"integer","format":"int64"},"nombreProducto":{"type":"string"},"descripcion":{"type":"string"},"precio":{"type":"number","format":"double"},"cantidad":{"type":"integer","format":"int32"},"estadoProducto":{"type":"string","enum":["DISPONIBLE","NO_DISPONIBLE"]},"categoria":{"$ref":"#/components/schemas/Categoria"}}}}}}